removed the lib services
